backend/requirements.txt
googleapis-common-protos==1.63.1
google-api-core==2.19.0
google-cloud-storage==2.17.0
rsa==4.9
Django==5.0.6
proto-plus==1.23.0
google-crc32c==1.5.0
requests==2.32.3
certifi==2024.6.2
charset-normalizer==3.3.2
django-cors-headers==4.3.1
protobuf==4.25.3
urllib3==2.2.2
asgiref==3.8.1
pyasn1==0.6.0
google-resumable-media==2.7.1
cachetools==5.3.3
google-cloud-core==2.4.1
sqlparse==0.5.0
pyasn1_modules==0.4.0
djangorestframework==3.15.2
idna==3.7
google-auth==2.30.0
python-decouple==3.8
